rename all of the 'transform*' methods to 'build*' methods, rename transform.go to builder.go
