+++ internal/compiler/compiler.go
@@ -18,7 +18,7 @@ func Compile(source parser.Source) CompilerOutput {
 	p1 := parser.NewParser(source)
 	escMod := p1.ParseModule()
 	builder := &codegen.Builder{}
-	jsMod := builder.TransformModule(escMod)
+	jsMod := builder.BuildModule(escMod)
 
 	p2 := codegen.NewPrinter()
 	p2.PrintModule(jsMod)
